Tweaks to xattrs for consistency.  Added FUSE object-type.
Alfred noticed some issues in the way we were generating xattr values

o  object-IDs weren't getting inodes in them

o  init_pre() and str_2_pre() had inconsistent formats for obj-ID

o  object-IDs weren't getting indicator for PACKED versus non-PACKED

This last one poses a slight conundrum.  We already encode the object-type
into the Post xattr.  (It waits until the Post xattr, because fuse doesn't
know whether an object is Uni or Multi until close-time.)  However,
recovery wants to be able to tell from the object-ID whether this is a
packed object or not.

The compromise is to add an object-type FUSE.  If you are looking at the
MDFS xattrs, you wont see this, because the Post xattr still has the
correct type.  If pftool is generating the object-ID, it can always supply
the correct type into the object-ID field (which may or may not be PACKED).
However, fuse can supply FUSE as the object-type, for object-ID purposes,
which is understood to be "maybe Uni or Multi, but definitely not PACKED".

Now you can look at object-IDs and determine whether they are PACKED, or
not.