MarFS open/write/close writes objects, using new streaming extensions to libaws4c.
A new test in the unit-tests in aws4c (test_aws.c, case 12) shows a new
capability for streaming PUTs, which we adapted a little for MarFS fuse.
The new support is implemented in object_stream.{c,h}
Here's a quick outline:

(a) open()

The FileHandle initialized in MarFS open now also includes an ObjectStream,
which holds all the state related to streaming objects.  We install a
custom readfunc into the IOBuf.  Then, a thread is spun-off to make the
call to s3_put().  This allows open() to return before the transaction is
complete.  The call to s3_put() causes curl to invoke the readfunc (to
"read" more data for curl), which blocks on a semaphore, waiting for data.

(b) write()

In the ObjectStream in the FileHandle, MarFS write() can access the
semaphores that control interaction with the streaming_readfunc described
in (a).  It also sees the IOBuf used by the readfun.  It "appends" the
caller's buf onto the IOBuf, allows the readfunc to proceed, and waits for
the readfunc to finish.  Curl typically calls the readfunc with room for
16k (or less) at-a-time.  The readfunc iterates until the MarFS write-buf
is all sent, then signals write().  Write() then returns.

NOTE: This implies a "synchronous" write, because otherwise we'd have to
copy the caller's buffer before returning from write.  This way, we avoid
the copy, but we have to wait for curl.

(c) close()

This is like write, but we send an empty buffer.  The readfunc recognizes
this as EOF, and returns 0 to curl, which indicates EOF to curl, and the
PUT is "closed".  The s3_put thread returns.

This all is just the very beginning of support for write().  We do
currently also write MarFS system xattrs onto the MDFS file, but we are not
yet handling Uni/Multi objects, or anything like that.