First-cut at streaming object-reads.  Can now write/read simple Marfs files.
(Previous work supported object-writes.)

Combed through return values to try to get systematic about returning 0 for
success, and -1 plus err-code for failure, with fuse returning negative
errcode.

TBD: Handle reading from a non-zero offset, and/or for a length that
doesn't end at EOF.

Moved logging to a separate file, so it can be used anywhere, and maybe
shared with libaws4c.