Improved trashing, date-strings, GC support, and closing of object-streams
Trash now includes support for the "truncate" case, where the original file
is copied, with all xattrs, to trash, and then the original is stripped of
all xattrs.  Renamed trash_file() to trash_unlink(), and trash_dup_file()
to trash_truncate(), for clarity.

Ancient glibc 2.12 (on rrz, caddy, test-bed) provides strptime/strftime
that have problems with time-zones and DST.  Could fix by using "%Z" for
timezones, because that includes DST spec, as in "MDT".  But only strptime
supports "%Z" in this crufty old version.  Instead, we now add an explicit
DST flag to the printed-rep of date-times.  Functions to handle the
stringification were abstracted out, so they could be used everywhere:
epoch_to_str() / str_to_epoch().

Added the "gc" field to the Post xattr.  This is empty for all MDFS files
that are not in the trash.  For files that are in trash, this field
contains the full path to the file in the trash.  The point is to allow a
GPFS ILM scan task (i.e. one that's implementing garbage collection) to get
everything it needs from the scan-values.  This allows it to see which
inodes are in the trash, how old they are, and what is the path to the file
to be deleted.

stream_sync() now attempts to avoid any possible deadlocks with ongoing
stream_read() or stream_write() operations.