Bug fix: mknod() should install xattr, so open() won't think MD file is DIRECT
The problem is that open() check for existence of any MarFS xattrs.  If the
file exists and has no xattrs, we assume the file was written at a time
when the repo.access_proto was PROTO_DIRECT.  That would mean it should be
treated as containing raw data, rather than being the usual MarFS
"meta-data"' file.

However, mknod() just creates an empty MD file with no xattrs.  You can see
the problem.  We solve this by just having mknod() install the RESTART
xattr (which was originally intended for use by pftool) ... unless
repo.access_proto really *is* PROTO_DIRECT, of course.

Now, open() sees xattrs, and properly treats file as an MD file.  And
close() removes the RESTART xattr.

BTW, installing RESTART this way also protects us from another problem: If
one process is writing object-IDs for a MULTI file into the MD file, but
hasn't closed it yet (when we currently install all xattrs), then someone
else comes along and reads it, they would've treated it as a DIRECT file,
and retreived the object-ID info that we were writing into the file.