Symlinks working.
There is a special consideration for the case of deleting a symlink, when
the link-destination is another marfs file (under the same mount-point).
In this case, calling access() from within unlink() attempts to follow the
link, and then do something that requires fuse (e.g. getattr()).  However,
we are already processing unlink(), so waiting for ourselves to respond to
the secondary request creates a deadlock.  (I thought fuse had multiple
threads for these cases, but apparently not.)  marfs_unlink() detcts this
case, and (in this case only) skips the call to access(), thus avoiding the
deadlock.