Support for Multi, including chunk-info in MDFS.  Prototype recovery-info.
(1) Fuse can now read/write Multi objects, including chunk-info written
into the MDFS file.  Fuse doesn't actually need this info, but we write it
to be compatible with pftool, which will write this to support restart.  We
write the chunk-info as binary data (in network-byte-order), because that's
what pftool will want to do, in order to allow rapid seeking and reading.

All object-names are the same as the original object-ID, but just changing
the chunk-number.  So, these don't need to be written as part of the
chunk-info.  Instead, we just save some info like data-offset (i.e. logical
offset of data starting at the beginning of this chunk), checksum, length
of user-data, etc.

(2) We also write recovery-info into the tail of each object.  For now,
this is just crap, but it is crap of the correct size, to match
RecoveryInfo, plus a final 8-byte value which can be either the offset of
the beginning of the recovery info, or the length of the recovery-info.
The point is that, should the entire MDFS be lost, you could rebuild it
(albeit painfully and imperfectly) by reading teh recovery-info in the
objects themselves.  This will eventually contain the original MDFS path
(at the time of creation), etc.  We will not maintain this info after the
object has been written, so moving a marfs file makes the MDFS path in the
recovery-info obsolete, though it may still be useful.

The presence of recovery=info in objects means that reading of Multi
objects must be aware of the difference between the "logical offset"
(i.e. the byte-index in the user's data), and the "physical offset"
(i.e. the number of bytes between offset-0 in object-0 a given offset in a
given object, counting all the recovery-info).  This appears to be working
correctly.

(3) Fixed a bug in unlink().  We use access() to determined whether a user
is allowed to unlink a file.  If the file is a symlink, and the destination
is within the same MarFS mount-point, then fuse will deadlock.  We avoid
this by testing for this situation and skipping the access(0 call in that
case.  This seems safe, on the idea that the unlink will simply fail, if
the access() would've failed.

(4) libaws4c moved the enable for chunked-transfer-encoding from
IOBuf.flags to AWSContext.  This was because of our issue with wiping this
out during reads of multi-objects, in which a stream is closed and reopened
with a new chunk.  (It would've showed up elsewhere, as well.)  Fixed in
libaws4c, with corresponding changes here.