Read PACKED files, support xattrs, change "slave" to "shard".
This implements a first-cut at reading PACKED files.  The idea is that
object-storage will be inefficient for small-sized files.  Therefore, the
object-stoage for each of many small files are concatenated together into a
single large object.  The offsets of the data for the member objects are
kept in their own Post xattrs.  To access the contents of any of the
original members, we use it's offset and length.  We are assuming that each
member object is followed by its RecoveryInfo record, plus 8 bytes.  This
is just like the format the member-object would have if it were stored as a
UNI file.  In fact it may orginally have been a UNI file.

Temporarily checking-in the script I used to build dummy packed files.
Fuse never writes such files (because we never know how big something is
going to be until it is closed), but it does have to read them.  This
script captures the current assumptions about how PACKED files are built.

Fuse also now supports setxattr/getxattr/listxattr.  setxattr refuses to
set anything that looks like a system-attribute, and getxattr refuses to
return values for such things, and listxattr strips system-keywords out of
the returned list.

We also changed the name for directory-shards from "slaves" to "shards".