Support repos using Scality sproxyd.
URLs had to be rearranged to support sproxyd.  In sproxyd, URLs must look
like this:

  http[s]://host:port/proxy/alias/etc

where:

  "proxy" is the fastcgi path, configured in
  /etc/httpd/conf.d/_scality/sproxyd.conf on the server.  This just
  redirects to a listening sproxy daemon, and ...

  "alias" is an sproxyd-driver-alias, configured in /etc/sproxyd.conf on
  the server.

So, we now generate URLs that look like this:

  http[s]://host:port/repo/namespace/etc

where:

  "repo" is Repo.name from the config, and

  "namespace" is encode_namespace(Namespace.mnt_suffix)

For S3, we now expect bucket-names to just match Repo, instead of what used
to be "<repo>.<namespace>"

For sproxyd, we expect the "proxy" name to match the Repo, and we expect
there to be an sproxyd driver-alias matching the de-slashified namespace.

If you define TRY_SPROXYD (e.g. 'make TRY_SPROXYD=1'), then fuse will set
up it's hard-wired config to use a Scality sproxyd repo.  Otherwise, we use
an EMC repo supporting S3.

libaws4c allows suppression of the "Authorization: AWS ..." header. These
headers are needed in curl requests to S3 servers, but not sproxyd servers.
Sproxyd simply ignores such headers, so they don't technically need to be
suppressed, but we want fuse to possibly run where ONLY sproxyd is
supported, and no ~/.awsAuth file exists, containing user-keys.  Therefore,
we're currently allowing a failure of the load of the libaws4c config-file
(~/.awsAuth) to be non-fatal, in the case where TRY_SPROXYD is defined.

TESTING: Tested for one MarFS file, using either EMC S3, or Scality sproxyd.