unlink() does copy.  "path" file fixed.  Inode-reps to decimal.  Test-build for Brett.
We've decided it isn't reasonable to assume unlink() can always be
implemented internally with a rename-to-trash, because we want to allow
trash to live in a fileset or filesystem that is different from the one
where meta-data lives.  We don't even try to rename, first.

I was writing the path to the trashed file into the "companion" trash file,
(named "<blah>.path") Should have been writing the "MarFS path" there, so
it can be properly restored by undelete.

Printed-representation of inode-numbers (e.g. in object-IDs) has been in
hex.  Dave points out that admins will want to use 'ls -i' and figure out
where a file would be stored into the trash (once we get to the
hierarchical set of directory-hashes, which hasn't been done yet.)  That
means the representation in the object-IDs, etc, should probably be in
decmial, for now.  So they are.

Added another hard-wired config.  This one supports Brettk, beginning to do
unit-tests.