Hardwired config now holds multiple repos and namespaces.
Previously, the hardwired config always had one repo and one namespace.  We
added special defines and Makefile environment-settings to custom-build so
that this repo/namespace could take on different values for different
purposes.  We want different users to have different namespaces, so they
don't use the same MDFS (or maybe object store), which would confuse
testing.  But, the growing number of special-purpose builds was starting to
get ugly and compilcated.

Soon (when config-files are supported), we're going to have multiple
namespaces and repos.  Might as well start supporting that now.  So, we now
support multiple namespaces and repos, and the "hardwired" config just
installs all the cases that used to require custom-builds.

There are namespaces for Alfred's GC testing, Brett's unit-testing, and
Jeff's various fuse testing.  We also have repos with different features.
For example, Brett wants a small chunksize for unit-testing, to allow
creating MULTI files without having to write huge amounts of data.  We have
repos that support Scality sproxyd, and others using EMC S3.

TBD: We also added a "root" namespace, to cover the case of
directoy-listings in the top-level marfs mount, and to allow us to respond
to OS statvfs calls on '/', which happen from time to time.  However, this
reveals a bunch of issues we have to think about, such as: what does 'stat'
of the top-level dir return?  Do we have to have special-cases everywhere
to handle the root NS differently?  How about controlling user-access to
this directory? (Done thorugh meta-permissions?)  Etc.