Reworked makefile.  Maintain POST.chunks.  Tighter matching for root NS.
Instead of expecting users to use a bunch of confusing
environment-variables to control different features of the build, we've now
just got different targets.  By default, everything deals with the
"standard" build, which compiles with '-O3', and logs to syslog (but
without curl conversation details).  There's also a "lean" build (no
logging at all), and a "dbg" build (compile '-g' and run in the foreground,
i.e. without threads).

If you run 'make' it will build the standard build.
If you run 'make install', it will mount after building.

POST.chunks was not being maintained for Multi files.  I think I've fixed
that, now.

As of the previous commit, we have multiple namespaces, including the
"root" namespace.  The root namespace basically represents the mount-point,
and will allow directory listings there.  Still working out some issues
with properties of the root NS.  But one thing we have to deal with now is
that it is a prefix of all the other namespaces.  Therefore,
find_namespace_by_path() should only match it if the given path matches the
root mnt-suffix exactly.  This avoid spurious matches of root, when
searching for a non-existing namespace.