Tweaks to set up first multi-repo/multi-namespace mount.
Improvements to makefile, to allow default build and mount to write curl
conversations to ./fuse.log and other diagnostics to syslog.  (Can't send
it all to syslog because curl diagnostics just go to stdout.)  This should
allow me to track down problems, if given a to-the-second timestamp when
the problem happened.

Tweaked the hardwired repos to match Scality/EMC servers, spreading the
load out to multiple targets.  Validated all of them (except the XFS
target) with a simple write/read test.