Fixed readdir crashing because opendir hadn't recognized failure.
Added TRY_GT0(), which is what marfs_opendir() should have been using
to test the return value from opendir().  Now, when running as a
non-root user, and attempting 'ls' on a marfs-directory corresponding
to an MDFS directory to which you don't have access, you'll simply get
a permission denied, instead of crashing fuse.