Moved md_path to POST xattr.  Integrate recent HTTPS tweaks.
We realized that we always want to be able to find the path to the current
location of a metadata file using only xattrs.  This allows generic fast
access via GPFS inode scans.  Our former practice was to install the path
in POST only when the file was moved to the trash, and to use the presence
of this path in POST as an indication that the file was trashed.

Now, we have a new field in POST containing flag-bits, one of which indicates
trash.  But the md_path is always in the POST xattr, so generic inode scans can
always have easy access.

In the case of trashed files, there is still a companion-file, but this
file no longer has xattrs, because we don't need the trash-seeker to be
able to find that file (which would be confusing).  Instead, the GC tool
would just find the trash files, getting the md_path from their xattrs, and
can then generate the name of the companion-file.  The contents of the
companion file are the MD-path of the trashed file before it was trashed.
This would allow an "undelete" task to restore the file.

Recent changes to libaws4c apparently allow us to correctly generate https
calls through S3.  Minor updates here, to match changed interfaces.

Tweaked make_packed.