Fix ftruncate.  Fix unlink not moving to trash.  Put md_path xattr only on some files.
(a) When overwriting files, kernal calls fuse with open() / ftrunate() /
close().  However our open() opens a stream to the object, and opens the
MDFS file for write.  Thus, ftruncate() needed new functionality to allow
an open object-stream to be aborted, such that the stream would not be
persisted.  Then it can abort the stream to the object belonging to the old
file, create a new updated object-ID, and open a new stream to that.  It
also needed to copy the MDFS contents out of the way, and call POSIX
ftruncate() to wipe the original metadata contents (if any).  Finally,
because we use stringified time_t values to represent md_ctime and
obj_ctime, the updated object-ID had the potential to be the same as the
old one, if the original write and the truncate happened within the same
second.  Added a field to object-IDs to assure uniqueness in this case.

(b) Added check_trunc to validate the actions of truncate.

(c) Many fuse functions only need a temporary static PathInfo struct to use
as an argument to expand_path_info(), so they can get the metadata path, do
some simple thing, and return.  There was no need to wipe the contents of
the PathInfo, for this.  But other functions that need a temporary PathInfo
struct for expand_path_info() may also do somewhat more.  For example,
unlink was failing to move to the trash, because a flag in the
garbage-filled PathInfo indicated that stat_xattrs had already been called,
whereas the detected-xattrs flags showed that none were found.  Now, we
always wipe the contents of temporary inline PathInfo structs, as we should
have been doing all along.

(d) Turns out, the idea of always putting md_path into xattrs (i.e. to
allow inode scans to get easy access) is not a good idea for anything
except trash and SEMI_DIRECT.  For other types, changes to directories
could require updating all the xattrs under the change, which is
impractical.  So, only save md_path into the xattrs for SEMI and trash.