Fix POST xattr w/ unknown obj-type.  Fix file-descriptors in CLOSE_WAIT.
The object-type field in POST xattrs weren't getting initialized properly,
in ftruncate. Fixed.

Truncate was also leaving 3 open file-descriptors.  These were the
read/write fds for moving metadata to the trash, and the object stream.  I
think we were always leaving object-streams open, for every write, but
hadn't noticed it until make_gc_data tried writing tons of files, and
started exceeding the limit of open descriptors.  The problem was that
aws_reset_connection() wasn't getting called in marfs_close(), but even if
it was, it was only setting things up to reset on the NEXT open.  But in
object-streams, every file gets its own connector.  Therefore, the reset
must act on the existing connection.  Tweaked libaws4c to support this.
and tweaked marfs_close() to use it.

Adding some scripts in fuse/scripts.  One of these is make_gc_data which
will allow Alfred to recreate directories with files and garbage, for
testing the GC and Quota scripts.