Fix stat_xattrs().  Makefile expects LIBAWS4C from env.  Changes in struct fields.
Unlink was failing in get_packed, because lgetxattr() in stat_xattrs() was
not writing a zero at the end of the field.  Thus, after reading the
'marfs_objid' xattr, when we read the 'marfs_post' xattr, is was appearing
that there were valid characters in the mdfs-path field, because of the
crap left in the buffer from the objid xattr value.  This got parsed into
the post.md_path field, where it overwrote the value installed by
expand_path_info(), causing 'file doesn't exist' on the next call to
lgetxattr().

Now that Alfred is also building and mounting a fuse daemon, we can no
longer hardwire the location of the libaws4c library, in the Makefile.  But
we don't yet have the config-file reader.  For the time being, we'll use
the environment-variable LIBAWS4C to point to the directory where the
library is found.