Updating make_packed, to generate simulated Packed files for testing.
We recently decided the "objs" component of the marfs_post xattr-value
should hold the number of objects in a packed file.  This will allow the GC
script to easily determine when the set of files in the trash that are all
part of the same packed object is a complete set, meaning that they can be
reaped, and the packed object can be deleted.

The size of RecoveryInfo has also recently changed, so this script needed
updating to get that right.

Fuse doesn't write packed files, but it does have to read them.  This
script also serves to create a dataset to test fuse reading of packed files.

The script uses a fuse-mount to create "realistic" small objects, which it
then concatenates into a single packed object, and adjusts xattrs of the
member objects, accordingly.