Fixed 'rm -rf /marfs/blah/subdir'
In this case, releasedir gets called to close a directory that no longer
exists.  The <path> it is given is "-".  So we just avoid calling
expand_path_info(), and check_perms(), in the case the path looks like
that.

TBD: Why are we calling expand_path_info() and check_perms() anyhow?