(1) Support "root" namespace (e.g. 'ls /marfs').  (2) Create dirs at start-up.
(1) The "root" namespace is the namespace '/'.  This is the "root" of the
filesystem under the "mnt_top" (see config file), where marfs is mounted.
This is needed for several kinds of things.  (a) the kernel periodically
calls 'stavfs <mount_point>', which comes to fuse as a statvfs request with
namespace '/'; (b) 'ls -l <mount_point>'.  For a non-root user, this is not
allowed.  For a root-user, it returns a list of all the namespaces, with
stat-info from the corresponding <mnt_suffix>; (c) 'mkdir -p
<mnt_point>/<ns>/a/b/c' was failing because the implicit stat of the
top-level parent requires expand_path_info() to stat the root namespace.

NOTE: We added R_META to the iperms of the "root" Namespace.  An admin
could disable the root namespace by removing this permission.

(2) We also attempt to create all the metadata directories needed for all
namespaces (TBD: and semi-direct repos).  We introduce a new format for the
trash directory which is a "scatter-tree", described in the project
documentation.  Basically, the config for a namespace specifies a
filesystem root, where trashed metadata goes.  These directories may
potentially hold many files (until GC can reap them), including old copies
of files that were overwritten.  In order to prevent any one directory from
having more files than the file-system will support, an MDFS file with
inode=xxxxxQRS is trashed in the directory
<trash>/<namespace>.<shard>/Q/R/S/. We also now create this directory-tree
at start-up.

We also create the <mdfs> directory mentioned for every namespace in the
config file.  And we touch the <fsinfo> file and trunc to 0, if it doesn't
already exist.  (This will become unnecessary, once the quota script gains
the power to read the config-file.  At that point, it can be the thing to
create the fsinfo files, and truncate them to proper size.)