(1) don't create top-level admin dirs, (2) put setuid bits off limits
(1) Ownership issues make it impossible to initialize missing top-level
files and directories (e.g. MDFS, trash, fsinfo).  Also, admins may want to
create corresponding file-sets, etc.  So, if any such files or directories
are missing, fuse will log an error an exit.  However, if the top-level
trash-dir does exist, but lacks the scatter-tree underneath it, we will
create that.

(2) We are planning to use the setuid bits to indicate when stat() must
read xattrs.  (For example, setuid on an MDFS file means it is semi-direct,
so the true file-size is computed by the underlying parallel FS, so stat
should read the xattr to find the location, and stat that.)  For now, we
just make it illegal for users (or root) to setuid through the fuse-mount.