Added "compatibility" to marfs_base.h, so people can try Brett's new config support.
Brett's new marfs_compatibility.{h,c} support is starting to use Ron's
config-parser.  However, this supercedes some of the macros, decls, enums,
and #defines that used to be in marfs_base.  I'm not ready to make the
transition yet, but others are starting to do that.  This commit allows you
to #include marfs_base.h in other applications that are using the new
configuration, while still building fuse (which uses the old static
configuration).

Just add this to your code:

   #define NEW_CONFIG
   #include "marfs_base.h"

WARNING: You will be depending on there consistency between the static
config in marfs_base.c and in the config-file loaded by Brett's code.

Fuse will use the old static config, but other applications that include
marfs_base.h can still work.