Allow unlink to work on symlinks, by just unlinking them.
We don't put xattrs on symlinks.  Furthermore, GPFS returns EPERM, from
lgetxattr on a symlink.  This was causing trouble for trash_unlink(), which
checks for xattrs.  The simplest fix was to fix stat_xattrs() to recognize
this quirk of GPFS.  The effect is that unlink now sees symlinks as
basically DIRECT things, which it unlinks without any moving to the trash.

This means undelete can not restore deleted symlinks, until we decide it's
worth the trouble to have trash_unlink() move them there.  For now, it
seems like this could cause trouble for the GC script, which expects xattrs
on all files in the trash.  Symlinks on GPFS can't have that, and what would
the xattr values be, anyhow?