Config-file had version 1.0, but existing files have version 0.1
For now, I've reverted the config-file to say version 0.1, so all the
existing files will not appear to be different from the loaded version.  At
some point, we can all agree to get rid of all old files, and then run GC,
and then change the version in the config-file.

While I was at it, I punted on teh old idea of using a float to hold the
configuration version.  It is always getting converted back and forth as
two ints, in order to be the "major" and "minor" version-numbers, then
reinstalled in the float.  Instead, we now just have a major and minor
version number, which are both uint16_t.