Added some features to the config-reading library
-- still using old patched version of Ron's code.

There is a #ifdef TBD in marfs_configuration.c showing changes to support
the newer configuration support, when we finish getting that working.

-- libconfig.a now includes functions for pretty-printing struct contents.

The contents of a Repo or Namespace, as text.  debug_namespace(), and
debug_repo().  These could be used as a quick debugging statement in your
code, but can also be used for a shell-script interface (see below).

-- Added marfs_config app, which can be called from shell-scripts

This allows scripts to get access to various parts of the config.  For
example:

    $ marfs_config -h

    Usage: marfs_config [options]

    options:
            -h                     help
            -r [ <repo_name> ]     show named Repo (or all repo-names)
            -n [ <ns_name> ]       show named NS   (or all NS-names)
            -t [ <setting-name> ]  show named top-level parm (or all)

    Can't use more than one of -r, -n, -t

So a script could do something like this:

    #!/bin/bash
    for N in `marfs_config -n 2> /dev/null`; do
       MDFS=`marfs_config -n $N | grep -i md_path | awk '{print $2}'`
       echo "NS $N has chunk-size $CHUNKSIZE"
    done

TBD: report top-level config-options, like mount-point, etc.