Overhaul of fuse read(), and friends, for performance.
The kernel can call fuse read() with any offset() and any length.  For
starters, we took the route of closing and openining a unique GET request
for every read call.  We had to accomodate the presence of recovery-info at
the tail of each individual objject.  When spanning across multiple objects
(in a MULTI), we of course required additional close/opens.  The
performance was terrible, but it allowed a first cut at functionality.

Now, we're taking the more-challenging path of keeping a GET stream open,
and assuming the reads will be sequential.  If they prove not to be
contiguous (e.g. because of a seek), only then do we close and reopen.  The
result is a 20x improvement in bandwidth, but somewhat-increased
complexity.