Updated to work with Ron's latest parser-support.
The flag -DTBD is used to isolate what seemed to be new requirements.  In
addition to that, some minor changes were made to the XML tags.  These are
reflected in the new config files.  We now require a distinct "blueprint"
config, which is used to generate the initial structures which then drive
the parser during its regular operation on the "regular" config.

   marfs_cctest_blueprint.cfg    [blueprint]
   marfs_cctest.cfp              [full configuration]

If you run 'make DEBUG=1' you'll generate libconfig.a, plus linked
applications, whic generate logging output (to stderr).  Otherwise,
the tools will run silently (expect for their intended output).

The tool 'marfs_config' can be used to examine the configuration in detail.
(See 'marfs-config -h' for help) We should start using this in scripts, to
get hold of the mount-point, MDFS directories, repo.chunksizes, etc.