Tweaks to the static-config for Chris DeJager
Changed to match recent freshly-minted GPFS filesets and namespaces
in the dynamic configuration, after we lost GPFS and built a new
filesystem, with filesets different from how they used to be.

The static config is not an exact match with the latest config, but should
be enough to let Chris keep going with pftool.