Fixed a bug in fuse_releasedir().  Performance improvement in fuse_write().
For chunked-transfer-encoding (CTE) streams, curl make callbacks with
buffers of size (16k - 12), to leave room for the 12 bytes of CTE header
overhead.  When fuse is called with 128k buffers, this allows 8 transfers
to curl at size 16k-12, plus one final transfer of 96 bytes, which is
inefficient.  In this case, we just do the 8 transfers, and tell the kernel
we moved 128k-96.  This buys us ~20% BW improvement.