Added optional spinlocks.  Fixed bug in inits for PUT.
(a) I was seeing ~200k context-switches/sec on a storage node, while
running concurrent writes through fuse.  This could be explained by the use
of semaphores to communicate between producer and consumer threads, to
coordinate fuse read/write with and curl callbacks for the corresponding
GET/PUT streams.

The spinlocks here are using the CMPXCHG x86 instruction to achieve
thread-safe locking without a system-call.

PoliteSpinLock calls sched_yield() repeatedly, while trying the lock, so
that it doesn't just hog the CPU.

Defining SPINLOCKS (e.g. 'make ... SPINLOCKS=1') causes us to use
PoliteSpinLock instead of semaphores, everywhere in object_stream.c

Unfortunately, it doesn't seem to be improving performance.  Therefore, I'm
leaving semaphores as the default mechanism.  However, spinlocks might turn
out to make a difference later, so I'm leaving the option in place.

(b) There was a bug in the initialization of the "empty"
semaphore/spinlock, in the case of PUTs.  This allowed marfs_write to
return before the thread had finished moving all data to curl.  That in
turn meant that the next marfs_write() might have the appearance of writing
non-contiguous data (i.e. an error), if the curl thread still hadn't
finished moving the previous data (and updating the record of total amount
moved) by then.