stream_open() takes an extra flags arg, so marfs_open() does, too.
This is the beginning of support for different needs in fuse versus pftool.
The open-flags allow pftool to avoid using chunked-transfer-encoding (CTE),
which should allow better bandwidth, and perhaps better co-existence of
threads.

fuse_write() never knows how much data is going to be sent before the
close, so it must use CTE, which it provides in the open-flags.  On the
other hand, pftool may know how much it is going to write, at open time.
By not giving the CTE flag, it can exploit this to write more-efficiently.

The open-flags are saved in the ObjectStream (in the MarFS_FileHandle), so
that, when streams must be closed/reopened (e.g. when write hits a
Multi-file size-boundary), they will simply keep the properties they
already had.  (This means that pftool can no longer write multi-files as a
single stream, unless it does so by writing with CTE, because the
"content-length" header provided at the head of the PUT stream will be
wrong if the whole object is not written to that stream.  However, it could
write each individual object as a complete stream, without CTE.)