(a) load-balancing across repo hosts.  (b) marfs_open() takes content-length.
(a) Commit 17b2c284 added a way to configure repos to allow load-balancing.
This makes good on that, in marfs_open().

(b) We also recently extended marfs_open() to add a stream_flags argument,
which could be used to turn chunked transfer-encoding on or off.  The idea
was that pending changes in stream_open() would allow addition of an
explicit "Content-Length: ..." HTTP header, with a length that would be set
by other means.

Now, it seems simpler for marfs_open() to ditch the flags argument, and
just take a length argument.  If the length is 0, we assume CTE.  If length
is non-zero, we assume an explicit content-length.