Finish support for pftool to write content-length for all members of a Multi.
stream_open() already allowed a length-argument, which (if non-zero) caused
an HTTP content-legth header to be sent, instead of using chunked
transfer-encoding (CTE).  And marfs_open() took the same argument, which it
could pass on.  But this capability wasn't yet well integrated with writing
Multi-type MarFS objects, where the open-length (e.g. the size of a source
file in pftool) most be broken up so that individual MarFS chunk objects
are each written with teh appropriate content-length.

Now, we do that.  We track the total open size in the write_status member
of the MarFS_FileHandle.  That way, when marfs_write() needs to close and
reopen object-streams, at the boundary of MarFS objects, it can include the
appropriate per-object content-length header.

This doesn't seem to make a big difference on the test-bed I'm using, where
I'm writing to sproxyd servers that are located directly on storage-hosts.
However, Chris Hoffman is writing to sproxyd servers on independent
connector-hosts, and he is seeing (a) with CTE, the connectors buffer
entire objects (e.g. 1GB), before passing them on to the sproxyd hosts,
then waiting for a reply, and (b) with content-lengths, the connectors
stream their input directly to output.  The CTE behavior adds e.g. 10
seconds of latency per object, which affects bandwidth, not to mention that
the buffering no doubt strains available memory.  (Okay, I mentioned it.)

NOTE: We are not yet handling N:1 writes (from pftool), where different
tasks are writing objects that are part of the same Multi-object.