Support for N:1 writes from pftool.
There's a new entry point: marfs_open_at_offset() [not used by fuse].  When
opening via this offset, you are allowed to pick a non-zero offset for your
writes, though the offset is required to fall at a chunk-boundary, of the
repo installed into the xattrs.  You take on responsbility that all your
writes will either be working toward a full-size chunk, or will be the last
chunk in a multi.  This is reflected in the FH_ALLOW_RISKY flag, which is
set in the MarFS_FileHandle, when opening with this interface.

The length provided to open_at_offset() represents all the user-data
(i.e. not counting the size of recovery info) that will be written through
this opening, and may be longer than an individual chunk.  We'll track it
across writes, closing and opening at chunk boundaries, as would be
expected.  Each chunk is opened with whichever is smaller, the user-data
portion of a full-sized chunk, or the residual part of the original
open-length.  We add to this length the size of the recovery info, which we
automatically append at the tail of each chunk.

There are coresponding extensions to marfs_write() and marfs_release(),
when writing and the FH_ALLOW_RISKY flag is set.  Notably, fuse never knows
when an opbject will be completed, so it writes with
chunked-transfer-encoding, and stops at object-boundaries to write
recovery-info.  Therefore, the release (i.e. close) must also write
recovery-info.  In teh case of pftool, we know exactly how much info to
write, and so write() also always adds the final recovery-info.  So, in the
case of pftool, release/close doesn't write recovery-info.

There are also functions batch_pre_process() and batch_post_process() which
should be called once (single-threaded) before opening, and after closing
respectively.  open_at_offset() won't see the full size of the final file,
so it can't choose the proper repository for the given MarFS namespace.
Instead, batch_pre_process() gets the full final size, and it can
initialize all the xattrs to their final state, before opening.  That
allows N:1 writes to use the proper repository, so they can find the
correct chunksize.