bug-fixing for pftool, libcurl 7.45, etc.
(1) Pftool had a problem because when a write via our stream_put()
times-out, it doesn't mean that data-movement has ceased, between the
readfunc and the user-buffers.  Therefore, cleanup must close the stream,
before freeing the buffer, or wiping the aws4c IOBuf in the file-handle.
This solves a problem where timeouts resulted in a segfault.

-- stream_sync() needs to be aware when streams had a timeout, so it can
   avoid interacting with them (e.g. writing recovery info).

-- stream_sync() should assure the thread is cancelled and joined, when
   there has been a timeout, so marfs_release() can avoid wiping the
   filehandle (including the aws4 IOBuf being accessed by the thread),
   until all thread activity is done.

We have two options:

(a) stream_put/get can use SAFE_WAIT_KILL(), which now assures the thread
is cancelled and joined right when the time-out happens.  It sets a flag to
indicate what it has done.  Then stream_sync() sees this flag and acts
accordingly, including avoiding interacting with the stream (e.g. to append
recovery-info).

(b) stream_put/get can use SAFE_WAIT(), which does not immediately kill the
thread, but does set a (different) flag.  Then stream_sync() can know that
it must do the cancel and join, while also avoiding further interaction
with the stream.

Both options are supported, and have been briefly tested.  Select either
one by your choice of SAFE_WAIT() or SAFE_WAIT_KILL(), in stream_put/get.
(a) is probably more robust, because it doesn't require good behavior (i.e.
closing before cleanup) from pftool, but (b) seems a better division of
labor, and pftool copy_file() has now been made to behave well.

(2) Our installed version of libcurl is 7.19.7.  We are experimenting with
a custom-built libcurl based on 7.45.0.  We notice that the latter doesn't
call streaming_readfunc() again, if stream_open() provided a content-length
(i.e. the request had a content-length header, rather than being
chunked-transfer-encoded), and the full number of chars matching the
content-length header has been sent.  In such a case, stream_sync()
shouldn't use "stream_put(..0)" to get the readfunc to quit, because the
readfunc will never see it (because no more callbacks), and we'll get a
deadlock, resulting in a timeout and failure, at close time for every
object-stream.  Fixed and tested.

(3) Also:

marfs_chmod() shouldn't negate err-codes.  (That's an old fuse-ism.)

marfs_open() stop printing the host we selected at random, for pftool
debugging.

Added the object-stream address to some of the logging output, so that when
many streams are operating concurrently, you can (sort of) distinguish the
activity in different threads.